requirements.txt
numpy
scikit-image
tensorflow-addons
opencv-python
tqdm
